back/requirements.txt
Flask
Flask-MySQL
PyMySQL